feat: more indicators, style: Image upscale (#4)
* classic Combat animations

redid the classic combat animations to remove pixilation

* female movement

upscaled female movement

* image work

working on upscaling some of the images that got pixilated from scaling

* cleanup ele

also adjusted female armor coloring

* invis and lev

* lev

* coloring adjustments

* ghost

* ghost shadow

* feat: more status indicators

indicators for
- lev
- invis
- silenced
- feared
- looting

you can launch with arugments and they will toggle on

example
`/lua run spritehud 120 female` - will open with size 120 and toggle your sprite to be a female

you can add any of the toggles on the config as an argument at load to have that set to on.

* upscale male

upscaled male sprites

added check for plate classes to default to plate combat animations and other non casters the noarmor animations.